Simplified the tuple prep scripts into one
